my-flask-app/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
python-dotenv==0.19.0
opencv-python==4.5.3.56
Pillow==8.3.1
numpy==1.21.0
face-recognition==1.3.0